requirements.txt
Flask
Flask-Login
PyJWT
Flask-Limiter
bcrypt
Flask-SQLAlchemy
